# WebAPI/wwwroot/swagger-original.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NotStopAlarm",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
    "/alarmclock": {
      "get": {
        "tags": [ "alarmclock" ],
        "summary": "Возвращает информацию о всех будильниках",
        "operationId": "getAlarmClock",
        "responses": {
          "200": {
            "description": "Будильники успешно возвращены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmClock"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmClock"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка синтаксиса"
          },
          "404": {
            "description": "Какие-либо будильники не найдены"
          }
        }
      },
      "post": {
        "tags": [ "alarmclock" ],
        "summary": "Создаёт новый будильник",
        "operationId": "addAlarmClock",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlarmClock"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/AlarmClock"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Будильник успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmClock"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmClock"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка синтаксиса"
          },
          "409": {
            "description": "Будильник на такие дату и время уже существует"
          }
        }
      }
    },
    "/alarmclock/{alarmClockTime}": {
      "get": {
        "tags": [ "alarmclock" ],
        "summary": "Возвращает будильник по заданным дате и времени",
        "operationId": "getAlarmClockByTime",
        "parameters": [
          {
            "name": "alarmClockTime",
            "in": "path",
            "description": "Время искомого будильника",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "data-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Будильник найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmClock"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmClock"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка синтаксиса"
          },
          "404": {
            "description": "Будильник не найден"
          }
        }
      },
      "put": {
        "tags": [ "alarmclock" ],
        "summary": "Обновляет существующий будильник по дате и времени",
        "operationId": "updateAlarmClock",
        "parameters": [
          {
            "name": "alarmClockTime",
            "in": "path",
            "description": "Время обновляемого будильника",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "data-time"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlarmClock"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/AlarmClock"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Существующий будильник изменён",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmClock"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmClock"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка синтаксиса"
          },
          "404": {
            "description": "Будильник не найден"
          }
        }
      },
      "delete": {
        "tags": [ "alarmclock" ],
        "summary": "Удаляет будильник по времени и дате",
        "operationId": "deleteAlarmClock",
        "parameters": [
          {
            "name": "alarmClockTime",
            "in": "path",
            "description": "Время искомого будильника",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "data-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Будильник успешно удалён"
          },
          "400": {
            "description": "Ошибка синтаксиса"
          },
          "404": {
            "description": "Будильник не найден"
          }
        }
      }
    },
    "/alarmclock/{alarmClockTime}/invertWork": {
      "patch": {
        "tags": [ "alarmclock" ],
        "summary": "Инвертирует работу будильника",
        "description": "Если будильник был выключен, то включает его. Если будильник был включён - выключает.",
        "operationId": "invertAlarmClock",
        "parameters": [
          {
            "name": "alarmClockTime",
            "in": "path",
            "description": "Время искомого будильника",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "data-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Работа будильника успешно инвертирована",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmClock"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/AlarmClock"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка синтаксиса"
          },
          "404": {
            "description": "Будильник не найден"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlarmClock": {
        "required": [ "alarmClockColor", "alarmTime", "isWorking", "name" ],
        "type": "object",
        "properties": {
          "alarmTime": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string",
            "example": "alarmClock"
          },
          "alarmClockColor": {
            "type": "string",
            "example": "green"
          },
          "isWorking": {
            "type": "boolean",
            "example": true
          }
        },
        "example": {
          "alarmTime": "2000-01-23T04:56:07.000",
          "alarmClockColor": "green",
          "name": "alarmClock",
          "isWorking": true
        },
        "xml": {
          "name": "alarmclock"
        }
      },
...
    }
  }
}